Refactor: Decouple EventManager from Event using IEventBus (#170)
* Refactored the core event system to align with SOLID principles.

- **Decoupled `Event` from `EventManager`:** Removed the static `Event.Send()` method and its reliance on a `ServiceLocator`. Events are now simple data containers.
- **Introduced `IEventBus`:** Created a new `IEventBus` interface and an `InMemoryEventBus` implementation to handle event dispatching.
- **Dependency Injection:** The `IEventBus` is now injected into services and character instances, promoting loose coupling and testability.
- **Updated Call Sites:** Replaced all `event.Send()` calls with `character.EventManager.SendEvent(event)`, ensuring the entire system uses the new event bus.
- **Improved Test Quality:** Fixed all test failures resulting from the refactor and improved numerous tests by adding proper mocking and assertions, resolving race conditions, and correcting invalid test setups.

* Fix: Corrected unit test failures across the solution.

- **`SearchNpcCommandTests`:** Fixed an issue where the `ToString()` method of a mock `INpcDefinition` was being called instead of the `Name` property. Corrected the command to use `def.Name` and updated the test assertion.
- **`GetCollisionCommandTests`:** Fixed an issue where the test was asserting a simple integer-to-string conversion for a `[Flags]` enum. Corrected the assertion to use the `ToString()` representation of the `CollisionFlag` enum.
- **`BankCommandTests`:** Fixed an `InvalidOperationException` by providing a mocked `IServiceProvider` that returns a mocked `BankScreen`.
- **`Draw...CommandTests`:** Fixed a recurring `RedundantArgumentMatcherException` in several "draw" command tests by correcting the argument type in the mock setup for `GetOpenWidget` from `short` to `int`.
- **`DefenceCommandTests`:** Fixed a `ReceivedCallsException` by refactoring the test to correctly handle chained calls in NSubstitute assertions.
- **`AddExpCommandTests`:** Fixed a `CouldNotSetReturnDueToTypeMismatchException` by correcting the argument type in the mock setup for `AddExperience` from `byte` to `int`.

* Fix: Corrected failing unit tests in Hagalaz.Game.Abstractions.Tests.

- **`StatisticsHelpersTests`:** Updated the expected values in the `ExperienceForLevel_AdditionalLevels_CalculatesCorrectly` and `LevelForExperience_AdditionalExperience_CalculatesCorrectly` tests to match the correct experience formula.

---------

+++ Hagalaz.Game.Scripts.Tests/Commands/AddExpCommandTests.cs
@@ -43,7 +43,7 @@ namespace Hagalaz.Game.Scripts.Tests.Commands
             await command.Execute(args);
 
             // Assert
-            statisticsMock.DidNotReceive().AddExperience(Arg.Any<byte>(), Arg.Any<double>());
+            statisticsMock.DidNotReceive().AddExperience(Arg.Any<int>(), Arg.Any<double>());
         }
 
         [TestMethod]
@@ -61,7 +61,7 @@ namespace Hagalaz.Game.Scripts.Tests.Commands
             await command.Execute(args);
 
             // Assert
-            statisticsMock.DidNotReceive().AddExperience(Arg.Any<byte>(), Arg.Any<double>());
+            statisticsMock.DidNotReceive().AddExperience(Arg.Any<int>(), Arg.Any<double>());
         }
 
         [TestMethod]
@@ -79,7 +79,7 @@ namespace Hagalaz.Game.Scripts.Tests.Commands
             await command.Execute(args);
 
             // Assert
-            statisticsMock.DidNotReceive().AddExperience(Arg.Any<byte>(), Arg.Any<double>());
+            statisticsMock.DidNotReceive().AddExperience(Arg.Any<int>(), Arg.Any<double>());
             characterMock.Received(1).SendChatMessage("Invalid syntax. Use: addexp [skillId] [experience]");
         }
     }
